requirements.txt
openpyxl>=3.1.2
xlrd>= 2.0.1
flask>=3.0.0
Pillow>=10.1.0
fdb>=2.0.2
pandas>=2.1.3
numpy>=1.26.1